requirements.txt
python-telegram-bot==13.14
sqlalchemy
pillow
toml
emoji==2.0.0
alembic
pyrogram==0.17.1
tgcrypto==1.2.3
numpy
